+++ Annex/src/utopia/annex/model/request/ApiRequest.scala
@@ -1,14 +1,14 @@
 package utopia.annex.model.request
 
 import utopia.access.http.Method
-import utopia.flow.generic.model.immutable.{Model, Value}
+import utopia.flow.generic.model.immutable.Value
 
 /**
   * Represents a relatively simple request that may be sent multiple times if need be
   * @author Mikko Hilpinen
   * @since 16.6.2020, v1
   */
-trait ApiRequest
+trait ApiRequest extends Retractable
 {
 	// ABSTRACT ----------------------------
 	
@@ -16,27 +16,20 @@ trait ApiRequest
 	  * @return Request method
 	  */
 	def method: Method
-	
 	/**
 	  * @return Request path (root path not included)
 	  */
 	def path: String
-	
 	/**
 	  * @return Request body value. Empty value if no body should be sent
 	  */
 	def body: Value
 	
-	/**
-	  * @return Whether this request has been deprecated and shouldn't be sent (anymore)
-	  */
-	def isDeprecated: Boolean
 	
-	/**
-	  * @return A model that can be stored locally to replicate this request in another session.
-	  *         None if this request needn't be persisted and replicated
-	  */
-	def persistingModel: Option[Model]
+	// COMPUTED ---------------------------
+	
+	@deprecated("Deprecated for removal. Renamed to .deprecated.", "v1.7")
+	def isDeprecated = deprecated
 	
 	
 	// IMPLEMENTED  -----------------------