+++ Scribe/Scribe-Client/Changes.md
@@ -1,6 +1,7 @@
 # Utopia Scribe Client - list of changes
 
 ## v1.0.1 (in development)
+Supports **Annex v1.7**
 ### Other changes
 - Scala version updated to 2.13.12